+++ static/screen/index.css
@@ -1,11 +1,13 @@
 #main_page{
-    background: url(https://upload.wikimedia.org/wikipedia/commons/e/e1/Logo_of_YouTube_%282015-2017%29.svg) no-repeat 50%;
     text-align: center;
-    height: 360px;
-    opacity: 0.5;
 }
 
 #container_search{
-    padding: 270px 100px 100px 100px;
+    padding: 20px 100px 20px 100px;
 }
 
+.btn-outline-success {
+    color: white;
+    border-color: grey;
+    background-color: grey;
+}
