+++ templates/category.html
@@ -6,15 +6,15 @@
 
 {% block content %}
 <div class="ranking_title">
-    <h3> 오늘의 인기 동영상 랭킹 TOP 50 </h3>
+
 </div>
 
 <div class="ranking_analysis">
 <div>
 <select  class="form-select" aria-label="Default select example" onchange="location = this.value;">
     <option selected>인기 동영상 분석</option>
-    <option value="/ranking/category">카테고리별 분석</option>
-    <option value="/ranking/channel">채널별 분석</option>
+    <option value="/latest_trend/category">카테고리별 분석</option>
+    <option value="/latest_trend/channel">채널별 분석</option>
 </select>
 </div>
 </div>
