+++ templates/channel_result.html
@@ -8,7 +8,7 @@
 
 <div class="dashboard">
 <div class="dashboard_left">
-<h5> {{ title }} 검색 결과 </h5>
+<h6> {{ title }} 검색 결과 </h6>
 
 <table id="table" class="table table-hover">
   <thead>
@@ -38,21 +38,21 @@
 
 <div class="dashboard_right">
 <div class="search">
-  <form id="search_channel" method="get" action="/search/search_result">
+  <form id="search_channel" method="get" action="/dashboard/channel_result">
   <input name="cha_name" type="text" class="form-control" placeholder="채널을 검색하세요." aria-label="채널을 검색하세요." aria-describedby="button-addon2">
   <input type="submit" value="검색">
   </form>
 </div>
 
 <div class="search">
-  <form id="search_channel_id" method="get" action="/search/video_result">
+  <form id="search_channel_id" method="get" action="/dashboard/video_result">
   <input name="channel_id" type="text" class="form-control" placeholder="채널아이디를 검색하세요." aria-label="채널아이디를 검색하세요." aria-describedby="button-addon2">
   <input type="submit" value="검색">
   </form>
 </div>
 
 <div class="search">
-  <form id="search_keyword" method="get" action="/search/filter_result">
+  <form id="search_keyword" method="get" action="/dashboard/filter_result">
   <input name="find" type="text" class="form-control" placeholder="키워드를 검색하세요." aria-label="키워드를 검색하세요." aria-describedby="button-addon2">
   <input type="submit" value="검색">
   </form>
