+++ templates/chart.html
@@ -6,18 +6,9 @@
 
 {% block content %}
 <div class="ranking_title">
-    <h3> 오늘의 인기 동영상 랭킹 TOP 50 </h3>
+
 </div>
 
-<div class="ranking_analysis">
-<div>
-<select class="form-select" aria-label="Default select example" onchange="if(this.value) location.href=(this.value);">
-    <option selected>인기 동영상 분석</option>
-    <option value="/ranking/category">카테고리별 분석</option>
-    <option value="/ranking/channel">채널별 분석</option>
-</select>
-</div>
-</div>
 
 <div class="ranking_table">
 <table id="table" class="table table-hover">
@@ -51,9 +42,10 @@
 {% endblock %}
 
 {% block scripts %}
+
   <script>
     $(document).ready(function () {
-      $('#table').dataTables({
+      $('#table').dataTable({
         columns: [
           {orderable: false, searchable: false},
           {orderable: false},
@@ -67,5 +59,7 @@
       });
     });
   </script>
+
+
 {% endblock %}
