+++ templates/trend.html
@@ -18,9 +18,4 @@
 </select>
 </div>
 </div>
-
-<div>
-    <div> 조회수 TOP5: {{ channeltitle }} </div>
-
-</div>
 {% endblock %}
